Still dirty, needs more linux source files?
CAN linux socket files are used and thus requires includes from kernel space. Thiis means that linux source code files need to be brought in for the project to make?

Upon researching, this is a bad practice as it gets ugly from the literal infinite chaining of headers needed. It's better to carve out what we need, and just implement that.